refactor(test-impact): Rename to impact-map term (no-changelog)
coverage-map.ts→impact-map.ts; public mergeCoverage→buildImpactMap; internal
helper (was buildImpactMap)→assembleImpactMap to avoid the collision. The
merge-coverage CLI command + aggregate-coverage caller unchanged. 68 tests +
janitor build green.
