+++ packages/@n8n/test-impact/src/select/strategy.ts
@@ -1,4 +1,4 @@
-import type { ChangedFile, ResolveResult } from '../coverage-map.js';
+import type { ChangedFile, ResolveResult } from '../impact-map.js';
 
 /**
  * A selection strategy: given changed files, decide which specs must run.
